# OpenAstroAra.Server/openapi.yaml
...
openapi: 3.1.0
info:
  title: OpenAstro Ara Server API
  version: 0.0.1-ara.1
  description: |
    REST + WebSocket API for the OpenAstro Ara headless daemon. Speaks to the
    Flutter WILMA client (`client/openastroara_client`). Designed for trusted-LAN
    deployments; v0.1.0 will add opt-in remote-access mode with TLS + token auth.

    **Lineage:** forked from N.I.N.A. (Nighttime Imaging 'N' Astronomy). See repo
    root LICENSE.txt, COPYING, AUTHORS for the inherited attribution. NOTICE.md
    (added at Phase 15) carries the full lineage record.
  license:
    name: MPL-2.0
    url: https://www.mozilla.org/en-US/MPL/2.0/
servers:
  - url: http://{host}:{port}/api/v1
    description: Direct LAN connection (default)
    variables:
      host:
        default: openastroara.local
        description: Pi hostname or IP; discoverable via mDNS service _openastroara._tcp.local
      port:
        default: "5555"

# ─── Endpoint group: Server ────────────────────────────────────────────────────
paths:
...
  /guiding/autotune/sessions/{sessionId}:
    get:
      tags: [Guiding Auto-Tune]
      summary: Get one persisted auto-tune session
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Session status
          content: { application/json: { schema: { $ref: '#/components/schemas/GuidingAutoTuneStatus' } } }
        '404': { description: Session not found }
  /guiding/autotune/sessions/{sessionId}/report:
    get:
      tags: [Guiding Auto-Tune]
      summary: Get one persisted auto-tune Markdown report
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Markdown report
          content: { application/json: { schema: { $ref: '#/components/schemas/GuidingAutoTuneReport' } } }
        '404': { description: Session not found }
  /guiding/autotune/sessions:
    post:
      tags: [Guiding Auto-Tune]
      summary: Analyze telemetry and optionally run bounded auto-tune experiments
      requestBody:
        required: false
        content: { application/json: { schema: { $ref: '#/components/schemas/GuidingAutoTuneStartRequest' } } }
      responses:
        '200':
          description: Proposal
          content: { application/json: { schema: { $ref: '#/components/schemas/GuidingAutoTuneStatus' } } }
        '400': { description: Invalid tuning bounds }
        '409': { description: Missing guider, telemetry, mount, or image scale }
...
  /guiding/autotune/sessions/{sessionId}/cancel:
    post:
      tags: [Guiding Auto-Tune]
      summary: Cancel the current auto-tune session by ID
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Restored status }
        '404': { description: Session not found or not current }
  /guiding/autotune/sessions/{sessionId}/apply:
    post:
      tags: [Guiding Auto-Tune]
      summary: Apply the current proposed settings by session ID
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Applied status }
        '404': { description: Session not found or not current }
  /guiding/autotune/sessions/{sessionId}/rollback:
    post:
      tags: [Guiding Auto-Tune]
      summary: Restore the current session snapshot by ID
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Restored status }
        '404': { description: Session not found or not current }

# ─── Reusable parameters + schemas ────────────────────────────────────────────
components:
...
  schemas:
    GuidingAutoTuneStartRequest:
      type: object
      properties:
        depth: { type: string, enum: [quick, standard, deep], default: standard }
        guide_pixel_scale_arcsec_per_pixel: { type: number, minimum: 0 }
        main_pixel_scale_arcsec_per_pixel: { type: number, minimum: 0 }
        guide_rate_arcsec_per_second: { type: number, exclusiveMinimum: 0, default: 7.5205 }
        allow_guide_rate_changes: { type: boolean, default: false }
        allow_algorithm_changes: { type: boolean, default: false }
        apply_automatically: { type: boolean, default: false }
        dry_run: { type: boolean, default: true }
        maximum_samples: { type: integer, minimum: 8, maximum: 12000, default: 4000 }
        maximum_characterization_seconds: { type: integer, minimum: 10, maximum: 2400, default: 0 }
        evaluation_seconds: { type: integer, minimum: 10, maximum: 1800, default: 180 }
        stabilization_seconds: { type: integer, minimum: 1, maximum: 300, default: 15 }
        use_main_camera_validation: { type: boolean, default: false }
        main_camera_validation_exposure_seconds: { type: number, exclusiveMinimum: 0, default: 1 }
        main_camera_validation_binning: { type: integer, minimum: 1, default: 1 }
        main_camera_validation_frames: { type: integer, minimum: 1, maximum: 8, default: 3 }
        maximum_candidates: { type: integer, minimum: 1, maximum: 12, default: 8 }
        maximum_session_minutes: { type: integer, minimum: 1, maximum: 180, default: 30 }
        minimum_altitude_degrees: { type: number, minimum: -90, maximum: 90, default: 45 }
...
    GuidingAutoTuneStatus:
      type: object
      required: [session_id, state, progress, current_step, telemetry_samples, can_apply, can_rollback, warnings, started_at_utc, updated_at_utc]
      properties:
        session_id: { type: string, format: uuid }
        state: { type: string }
        progress: { type: number, minimum: 0, maximum: 1 }
        current_step: { type: string }
        behavior_class: { type: string, nullable: true }
        behavior_confidence: { type: number, nullable: true }
        telemetry_samples: { type: integer }
        baseline_score: { type: number, nullable: true }
        best_score: { type: number, nullable: true }
        can_apply: { type: boolean }
        can_rollback: { type: boolean }
        warnings: { type: array, items: { type: string } }
        plan: { type: object, nullable: true }
        best_candidate: { type: object, nullable: true }
        baseline_result: { type: object, nullable: true }
        candidate_results: { type: array, items: { type: object }, nullable: true }
        started_at_utc: { type: string, format: date-time }
        updated_at_utc: { type: string, format: date-time }
    GuidingAutoTuneReport:
      type: object
      required: [session_id, content_type, markdown, generated_at_utc]
      properties:
        session_id: { type: string, format: uuid }
        content_type: { type: string, example: text/markdown }
        markdown: { type: string }
        generated_at_utc: { type: string, format: date-time }
...